Implement Phases 2-4: Project Planning, Productivity, and Advanced Features
Phase 2 - Project Planning:
- Task dependencies (blocks/blocked-by/related) with cycle detection
- Milestones with progress tracking and deadline management
- Calendar view for due date visualization
- Reports & analytics with SVG charts (burndown, velocity, priority)
- Gantt chart view with timeline bars and dependency lines

Phase 3 - Productivity:
- Recurring tasks with configurable patterns (daily/weekly/monthly/custom)
- Custom fields per board (text/number/date/dropdown/checkbox)
- Time tracking with start/stop timer and manual entries
- Bulk operations (move, assign priority, delete multiple tasks)
- Keyboard shortcuts with help overlay (?, n, /, 1-6 view modes)

Phase 4 - Advanced:
- Project templates (save board as template, create from template)
- Workflow automation (trigger-based rules with configurable actions)
- Import/export (CSV, JSON, Trello import)
- Client portal (shareable board links with password protection)
- Webhooks (configurable event-based HTTP notifications)

Each feature includes full backend (model + queries + routes) and frontend
(service + components) with proper auth, board membership verification,
and multi-tenant isolation.

+++ backend/crates/db/src/migrations/20260209000001_phase1_features.sql
@@ -41,7 +41,7 @@ CREATE TRIGGER update_subtasks_updated_at BEFORE UPDATE ON subtasks FOR EACH ROW
 -- ============================================
 
 ALTER TABLE tasks ADD COLUMN IF NOT EXISTS start_date TIMESTAMPTZ;
-ALTER TABLE tasks ADD COLUMN IF NOT EXISTS estimated_hours DECIMAL;
+ALTER TABLE tasks ADD COLUMN IF NOT EXISTS estimated_hours DOUBLE PRECISION;
 
 -- ============================================
 -- Invitation Extensions
